ci: another attempt to fix CI
