--- tests/mod.test.ts
@@ -1,6 +1,0 @@
-import "./01_main.ts";
-import "./02_guild.ts";
-import "./03_role.ts";
-import "./04_channel.ts";
-import "./05_message.ts";
-import "./06_cleanup.ts";
